Made a new grid for Marcos and Oscar's MAMEs
Marcos and Oscar should now use the 3_3 grid to separate the tiles.
Also, the grid I'm using is 2/3 the new tile. This means that before
using the sprites taken out with the 3_3 grid, we'll have to redimension
them to 2/3 its size.
Remember to use the approximation method in photoshop if you're
redimensioning the sprites. Otherwise, we won't get clean sprites.